delete old rules file
